pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wifi-cam-mcp"
version = "0.1.0"
description = "MCP server for controlling WiFi cameras (Tapo C210) - Let AI look around your room!"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pytapo>=3.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
wifi-cam-mcp = "wifi_cam_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/wifi_cam_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
